feat: use autogenerated mapper
